tests: datadog trace dict when env vars not set
Added a test which checks that the `env`, `service`, and `version`
fields are empty when the `DD_*` environment variables are not set.

This also moves the logic to set the `DD_*` environment variables and
reload `ddtrace` into the test that requires it only from the
`conftest.py` file. The test also unsets the environment variables, and
again reloads `ddtrace` to avoid test pollution.

<https://uktrade.atlassian.net/browse/DBTP-1662>