build: Specify GH Actions Workflow permissions for test (off-ticket) (#32)
Resolves this CodeQL warning:

```
If a GitHub Actions job or workflow has no explicit permissions set, then the repository
permissions are used. Repositories created under organizations inherit the organization
permissions. The organizations or repositories created before February 2023 have the
default permissions set to read-write. Often these permissions do not adhere to the
principle of least privilege and can be reduced to read-only, leaving the write permission
only to a specific types as issues: write or pull-requests: write.
```