feat: Enable Datadog Error Tracking for Logs
To manually Enable Error Tracking for Logs, the `error.kind` and one of
`error.message` or `error.stacktrace` attributes must be present in the
log.

From DLFA, we're only concerned about adding errors for logs which
aren't triggered by an exception, either handled or unhandled, as this
is handled by the `ddtrace-py` library.

This functionality is being added to mirror the automatic logger
integration provided by the Sentry SDK within Datadog.